pyproject.toml
[tool.poetry]
name = "arduino_weather"
version = "0.1.0"
description = ""
authors = ["Max Wittig <max.wittig@siemens.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pyserial = "^3.4"
pytz = "^2019.3"
pyowm = "^2.10.0"
loguru = "^0.4.1"

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
